feedback PR + moved cell_transition functions to mixins
