+++ moscot/problems/time/_lineage.py
@@ -186,107 +186,6 @@ class TemporalProblem(TemporalAnalysisMixin, SingleCompoundProblem):
             return result
         self._dict_to_adata(result, result_key)
 
-    def _validate_args_cell_transition(
-        self, arg: Union[str, Mapping[str, Sequence[Any]]]
-    ) -> Tuple[Union[str, Sequence], Sequence]:
-        if isinstance(arg, str):
-            if not hasattr(self.adata.obs[arg], "cat"):
-                raise ValueError(f"The column `{arg}` in `adata.obs` must be of categorical dtype")
-            return arg, list(self.adata.obs[arg].unique())
-        _key, _arg = arg.keys(), arg.values()
-        if not hasattr(self.adata.obs[_key], "cat"):
-            raise ValueError(f"The column `{_key}` in `adata.obs` must be of categorical dtype")
-        if len(_key) != 1:
-            raise ValueError("The data can only be filtered according to one column of `adata.obs`")
-        if not set(_arg).isin(set(self.adata.obs[_key].unique())):
-            raise ValueError(f"Not all values {_arg} could be found in column {_key}")
-        return _key, _arg
-
-    def cell_transition(
-        self,
-        start: Any,
-        end: Any,
-        early_cells: Union[str, Mapping[str, Sequence[Any]]],
-        late_cells: Union[str, Mapping[str, Sequence[Any]]],
-        forward: bool = False,  # return value will be row-stochastic if forward=True, else column-stochastic
-        **kwargs: Any,
-    ) -> pd.DataFrame:
-        _early_cells_key, _early_cells = self._validate_args_cell_transition(early_cells)
-        _late_cells_key, _late_cells = self._validate_args_cell_transition(late_cells)
-
-        transition_table = pd.DataFrame(
-            np.zeros((len(_early_cells), len(_late_cells))), index=_early_cells, columns=_late_cells
-        )
-
-        df_late = self.adata[self.adata.obs[self._temporal_key] == end].obs[[_late_cells_key]].copy()
-        df_early = self.adata[self.adata.obs[self._temporal_key] == start].obs[[_early_cells_key]].copy()
-
-        if forward:
-            _early_cells_present = set(_early_cells).intersection(set(df_early[_early_cells_key].unique()))
-            for subset in _early_cells:
-                if subset not in _early_cells_present:
-                    transition_table.loc[subset, :] = np.nan
-                    continue
-                try:
-                    result = self.push(
-                        start=start,
-                        end=end,
-                        data=_early_cells_key,
-                        subset=subset,
-                        normalize=True,
-                        return_all=False,
-                        scale_by_marginals=True,
-                        **kwargs,
-                    )
-                except ValueError as e:
-                    if "no mass" in str(e):  # TODO: adapt
-                        logging.info(
-                            f"No data points corresponding to {subset} found in `adata.obs[groups_key]` for {start}"
-                        )
-                        result = np.nan
-                    else:
-                        raise
-                df_late.loc[:, "distribution"] = result / np.sum(result)
-                target_cell_dist = df_late[df_late[_late_cells_key].isin(_late_cells)].groupby(_late_cells_key).sum()
-                transition_table.loc[subset, :] = [
-                    target_cell_dist.loc[cell_type, "distribution"]
-                    if cell_type in target_cell_dist.distribution.index
-                    else 0
-                    for cell_type in _late_cells
-                ]
-            return transition_table
-        _late_cells_present = set(_late_cells).intersection(set(df_late[_late_cells_key].unique()))
-        for subset in _late_cells:
-            if subset not in _late_cells_present:
-                transition_table.loc[:, subset] = np.nan
-                continue
-            try:
-                result = self.pull(
-                    start=start,
-                    end=end,
-                    data=_late_cells_key,
-                    subset=subset,
-                    normalize=True,
-                    return_all=False,
-                    scale_by_marginals=True,
-                    **kwargs,
-                )
-            except ValueError as e:
-                if "no mass" in str(e):  # TODO: adapt
-                    logging.info(f"No data points corresponding to {subset} found in `adata.obs[groups_key]` for {end}")
-                    result = np.nan
-                else:
-                    raise
-            df_early.loc[:, "distribution"] = result / np.sum(result)
-            target_cell_dist = df_early[df_early[_early_cells_key].isin(_early_cells)].groupby(_early_cells_key).sum()
-            transition_table.loc[:, subset] = [
-                target_cell_dist.loc[cell_type, "distribution"]
-                if cell_type in target_cell_dist.distribution.index
-                else 0
-                for cell_type in _early_cells
-            ]
-        return transition_table
-
     @property
     def growth_rates(self) -> pd.DataFrame:
         cols = [f"g_{i}" for i in range(list(self)[0][1].growth_rates.shape[1])]
